requirements/dev.txt
-r common.txt
django-debug-toolbar==1.4
flake8==2.5.1
tox==2.3.1
